Apply new header structure to unit.h
